Fix precision rendering bug with small slices
There was a visual rendering bug where if you had a small slice (e.g.
with a value of '1') then the arcs would appear as lines and the outline
of the circle would appear jagged. This was due to rounding of the X,Y
coordinates. This commit fixes that by removing the rounding.