Remove unnecessary size prop
This was leftover from a time when we were setting the width and height
on the SVG element. However, now we are only using it to set the viewBox
and do calculations based on the viewBox size. This is now all just
internal details so we can simplify by removing it entirely.