Update for React 0.14
The latest release of React has split react-dom and react-addon-*
packages out of the core react package. This, along with some other
deprecations from the previous upgrade caused us to have to make a
couple of adjustments to our tests. Other than that, this is a pretty
straightforward upgrade for us here.

While I was at it, I moved `react` from `dependencies` to
`devDependencies` to prevent multiple versions from being installed and
used by consumers of this project.