Convert to ES2015 class
I think I'd like to make this a stateless pure functional component, but
in order to fully do that I will need to change the way it is tested
some. That is more in depth than I'd like to go at the moment, so in the
meantime I'm taking this one step closer to that world by converting it
to an ES2015 class which requires some of the same changes that the pure
functional component would necessitate.

While I was doing this, I removed the className propType, which was
leftover from when we originally extracted this component from the
Brigade codebase but is no longer actually used.